Fix to include both splits in collision detection
Fix generate_universe to check for collisions of both newly split
bacterium.

Originally, when a split occured, the original bacterium was replaced by
one of the new split bacteria. The other new bacterium would be appended
to the end of the list. However, the collision matrix was not updated to
include that newly appended bacterium. As a result, the subsequent call
to collisionEventsHandler.run(), which adjusts the bacteria so they
don't overlap, ignored half of the newly split bacteria.

This should reduce overlap in frames with densely packed bacteria.

 - Add expand_collision_matrix to quickly update the collision matrix to
   include newly added bacteria.

 - Change in behavior of the algorithm if and only if a split occurs
   while the original bacterium is overlaping with other bacteria.