Cleaned up multiprocess-mapped functions and prepared for a bugfix.
The mphelper module was added to increase the readability and
maintainability of multiprocess-mapped functions found in the
helperMethods module. In addition, changes were made to the
expand_collision_matrix function in helperMethods to attempt to fix
a bug. However, the old behavior is still the current behavior and
those changes are currently unused.

- Added doublestar_function decorator to allow writing
multiprocess-mapped functions with multiple arguments.
- Added HandleExceptions static class to decorate multiprocess-
functions with a try-except statement that outputs the traceback
and exception to the console before the processes is terminated.
- Added kwargs to create dictionaries via a function call (just
to distinguish dictionaries from keyword arguments for better
readability).
- Modified expand_collision_matrix to actually expand the collision
matrix, which it wasn't actually doing. The change is currently
unused.